Extract indexers from SearchRepository (Phase 3)
Move document-building out of SearchRepository into dedicated, focused
indexers, leaving the repository responsible for AppSearch persistence and
orchestration:

- AppIndexer: launchable apps across profiles
- ShortcutIndexer: dynamic/manifest/pinned, static, and custom shortcuts
- ContactIndexer: contacts with phone/email search data
- SnippetIndexer: user-defined text snippets

Indexers are pure readers (retaining icon-disk-caching side effects). The
"abort on DeadObjectException" semantics of dynamic shortcut indexing are
preserved via a nullable return. Behavior unchanged; SearchRepository.kt
drops from 2,301 to 1,968 lines. ./gradlew testDebugUnitTest assembleDebug
passes.
